Added some logic to choose a to_string implementation. Fails on msvc due to a constexpr bug. compiles fine with llvm
Former-commit-id: 99656f56255fcf248e2b0cda3a5000a7fd5ee922
Former-commit-id: 28cd4ea7164f664887269606da956ff030801fa0
Former-commit-id: a49dd650927bdc50bea8c0f814a3a8762e0afb42

+++ Headers/Logger.h
@@ -28,7 +28,7 @@ public:
 	{
 #pragma warning( push )
 #pragma warning( disable : 6386 )
-		const size_t size = 512;
+		const int size = 512;
 		char* buffer = nullptr;
 		buffer = new char[size];
 		va_list vl;
@@ -60,7 +60,7 @@ public:
 	{
 #pragma warning( push )
 #pragma warning( disable : 6386 )
-		const size_t size = 512;
+		const int size = 512;
 		char* buffer = nullptr;
 		buffer = new char[size];
 		va_list vl;
@@ -84,7 +84,7 @@ public:
 	{
 #pragma warning( push )
 #pragma warning( disable : 6386 )
-		const size_t size = 512;
+		const int size = 512;
 		char* buffer = 0;;
 		buffer = new char[size];;
 		va_list vl;;
